Revert "use npm instead of exec, and other improvements"
