+++ hook-template
@@ -1,0 +1,19 @@
+#!/usr/bin/env node
+"use strict";
+var path = require('path');
+var execSync = require('child_process').execSync;
+var hook = path.basename(__filename);
+var config = require(path.resolve(process.cwd(), 'package'));
+
+if (!config.scripts[hook]) {
+  process.exit(0);
+}
+
+try {
+  var output = execSync(config.scripts[hook], { encoding: 'utf-8' });
+  console.log(output);
+} catch(error) {
+  console.log(error.stdout);
+  console.error(error.stderr);
+  process.exit(error.status);
+}
