+++ test/index.js
@@ -1,25 +1,19 @@
 "use strict";
-var fs = require('fs');
-var exec = require('child_process').exec;
+var fs = require('fs-extra');
 var path = require('path');
 var test = require('tape');
 
 test('creates hooks', function(is) {
-    is.plan(2);
+  is.plan(2);
 
-    fs.exists(__dirname + '/.git', function(exists) {
-        if (!exists) {
-            fs.mkdirSync(__dirname + '/.git');
-            fs.mkdirSync(__dirname + '/.git/hooks');
-        }
-    });
+  fs.removeSync('test/.git');
+  fs.ensureDirSync('test/.git/hooks');
 
-    exec('node ' + __dirname + '/../src/install', function(error, stdout, stderr) {
-        var hooks = fs.readdirSync(__dirname + '/.git/hooks');
-        var stat = fs.statSync(path.join(__dirname, '/.git/hooks', hooks[0]));
+  require('../');
 
-        is.equal(hooks.length, 4);
-        is.equal(stat.mode, 33261); // 755
-    });
+  var hooks = fs.readdirSync('test/.git/hooks');
+  var stat = fs.statSync(path.join(__dirname, '.git/hooks', hooks[0]));
 
+  is.equal(hooks.length, 4);
+  is.equal(stat.mode, 33261); // 755
 });
